+++ jupyterhub/handlers/pages.py
@@ -2,6 +2,7 @@
 # Copyright (c) Jupyter Development Team.
 # Distributed under the terms of the Modified BSD License.
 import asyncio
+import codecs
 import copy
 import time
 from collections import defaultdict
@@ -61,9 +62,9 @@ class HomeHandler(BaseHandler):
         # to establish that this is an explicit spawn request rather
         # than an implicit one, which can be caused by any link to `/user/:name(/:server_name)`
         if user.active:
-            url = url_path_join(self.base_url, 'user', user.name)
+            url = url_path_join(self.base_url, 'user', user.escaped_name)
         else:
-            url = url_path_join(self.hub.base_url, 'spawn', user.name)
+            url = url_path_join(self.hub.base_url, 'spawn', user.escaped_name)
 
         html = self.render_template(
             'home.html',
@@ -132,7 +133,7 @@ class SpawnHandler(BaseHandler):
         # which may get handled by the default server if they aren't ready yet
 
         pending_url = url_path_join(
-            self.hub.base_url, "spawn-pending", user.name, server_name
+            self.hub.base_url, "spawn-pending", user.escaped_name, server_name
         )
 
         if self.get_argument('next', None):
@@ -219,7 +220,7 @@ class SpawnHandler(BaseHandler):
         next_url = self.get_next_url(
             user,
             default=url_path_join(
-                self.hub.base_url, "spawn-pending", user.name, server_name
+                self.hub.base_url, "spawn-pending", user.escaped_name, server_name
             ),
         )
         self.redirect(next_url)
@@ -282,7 +283,9 @@ class SpawnPendingHandler(BaseHandler):
             # We should point the user to Home if the most recent spawn failed.
             exc = spawner._spawn_future.exception()
             self.log.error("Previous spawn for %s failed: %s", spawner._log_name, exc)
-            spawn_url = url_path_join(self.hub.base_url, "spawn", user.escaped_name)
+            spawn_url = url_path_join(
+                self.hub.base_url, "spawn", user.escaped_name, server_name
+            )
             self.set_status(500)
             html = self.render_template(
                 "not_running.html",
@@ -327,7 +330,9 @@ class SpawnPendingHandler(BaseHandler):
         # further, set status to 404 because this is not
         # serving the expected page
         if status is not None:
-            spawn_url = url_path_join(self.hub.base_url, "spawn", user.escaped_name)
+            spawn_url = url_path_join(
+                self.hub.base_url, "spawn", user.escaped_name, server_name
+            )
             html = self.render_template(
                 "not_running.html",
                 user=user,
@@ -348,6 +353,7 @@ class SpawnPendingHandler(BaseHandler):
 class AdminHandler(BaseHandler):
     """Render the admin page."""
 
+    @web.authenticated
     @admin_only
     def get(self):
         available = {'name', 'admin', 'running', 'last_activity'}