+++ jupyterhub/user.py
@@ -1,5 +1,6 @@
 # Copyright (c) Jupyter Development Team.
 # Distributed under the terms of the Modified BSD License.
+import json
 import warnings
 from collections import defaultdict
 from datetime import datetime
@@ -21,6 +22,7 @@ from .crypto import decrypt
 from .crypto import encrypt
 from .crypto import EncryptionUnavailable
 from .crypto import InvalidToken
+from .metrics import RUNNING_SERVERS
 from .metrics import TOTAL_USERS
 from .objects import Server
 from .spawner import LocalProcessSpawner
@@ -348,6 +350,11 @@ class User:
     def escaped_name(self):
         """My name, escaped for use in URLs, cookies, etc."""
         return quote(self.name, safe='@~')
+
+    @property
+    def json_escaped_name(self):
+        """The user name, escaped for use in javascript inserts, etc."""
+        return json.dumps(self.name)[1:-1]
 
     @property
     def proxy_spec(self):
@@ -753,6 +760,7 @@ class User:
                     self.db.delete(oauth_client)
             self.db.commit()
             self.log.debug("Finished stopping %s", spawner._log_name)
+            RUNNING_SERVERS.dec()
         finally:
             spawner.orm_spawner.started = None
             self.db.commit()